MLIBZ-2316: Handle skip and limit for Delta Set
We will log that the skip and limit modifiers will be ignored for a delta set request. These modifiers will be appliead to the cache after the delta set request has completed. A find() and pull() on a data store will now honor skip and limit modifiers preventing the user from keeping track of when they can and cannot use a specific query.