fix: do not require git when using the kinvey-nativescript-sdk
Currently, in order to use `kinvey-nativescript-sdk`, you need to have git installed on your machine. The problem is that one of the dependencies (pubnub) is declared with `git+https`.
Replace this with direct url to github, which is a .tgz file. This way you'll be able to install `kinvey-nativescript-sdk` without having `git` on the machine.