fix: update webpack configuration to polyfill Buffer
Webpack v5 no longer polyfill node.js APIs including Buffer.
Therefore added Buffer library polyfill and
configured Webpack to leverage the provided package as a “fallback” for APIs
which are not natively available in the target environment (browser).
The Webpack ProvidePlugin provides a mechanism for injecting a module/value
as a replacement/API for global variables